+++ src/ErrorHandler.php
@@ -14,6 +14,7 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
+use Psr\Log\LoggerInterface;
 use Throwable;
 
 class ErrorHandler implements MiddlewareInterface
@@ -21,13 +22,21 @@ class ErrorHandler implements MiddlewareInterface
     /** @var FormatterInterface[] */
     private $formatters = [];
 
+    /** @var LoggerInterface|null */
+    private $logger = null;
+
+    /** @var callable|null */
+    private $logCallback = null;
+
     /**
      * Configure the error formatters
      *
      * @param FormatterInterface[] $formatters
      */
-    public function __construct(?array $formatters = null)
+    public function __construct(?array $formatters = null, ?LoggerInterface $logger = null)
     {
+        $this->logger = $logger;
+
         if (empty($formatters)) {
             $formatters = [
                 new PlainFormatter(),
@@ -54,11 +63,33 @@ class ErrorHandler implements MiddlewareInterface
         return $this;
     }
 
+    /**
+     * @param callable $callback
+     */
+    public function logCallback(callable $callback): self
+    {
+        $this->logCallback = $callback;
+
+        return $this;
+    }
+
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
         try {
             return $handler->handle($request);
         } catch (Throwable $error) {
+            if ($this->logger) {
+                if ($this->logCallback) {
+                    ($this->logCallback)($this->logger, $error, $request);
+                } else {
+                    $this->logger->critical('Uncaught exception', [
+                        'message' => $error->getMessage(),
+                        'file' => $error->getFile(),
+                        'line' => $error->getLine(),
+                    ]);
+                }
+            }
+
             foreach ($this->formatters as $formatter) {
                 if ($formatter->isValid($error, $request)) {
                     return $formatter->handle($error, $request);